@@ -1,6 +1,8 @@
-//Created by Matthias Mueller - Intel Intelligent Systems Lab - 2020
+// Created by Matthias Mueller - Intel Intelligent Systems Lab - 2020
 
 package org.openbot.env;
+
+import static java.nio.charset.StandardCharsets.UTF_8;
 
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
@@ -12,35 +14,29 @@ import android.hardware.usb.UsbDeviceConnection;
 import android.hardware.usb.UsbManager;
 import android.os.AsyncTask;
 import android.widget.Toast;
-
 import androidx.localbroadcastmanager.content.LocalBroadcastManager;
-
 import com.felhr.usbserial.UsbSerialDevice;
 import com.felhr.usbserial.UsbSerialInterface;
-
 import java.io.UnsupportedEncodingException;
 import java.util.Map;
-
-import org.openbot.SensorService;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
+import org.openbot.common.Constants;
 
 public class UsbConnection {
-  private static final int USB_VENDOR_ID = 6790;  //0x2341; // 9025
-  private static final int USB_PRODUCT_ID = 29987;  //0x0001;
+  private static final int USB_VENDOR_ID = 6790; // 0x2341; // 9025
+  private static final int USB_PRODUCT_ID = 29987; // 0x0001;
   private static final Logger LOGGER = new Logger();
 
-  private UsbManager usbManager;
-  //private UsbDevice usbDevice;
+  private final UsbManager usbManager;
+  // private UsbDevice usbDevice;
   PendingIntent usbPermissionIntent;
   private static final String ACTION_USB_PERMISSION = "UsbConnection.USB_PERMISSION";
 
   private UsbDeviceConnection connection;
   private UsbSerialDevice serialDevice;
-  private LocalBroadcastManager mLocalBroadcastManager;
+  private final LocalBroadcastManager localBroadcastManager;
   private String buffer = "";
-  private Context context;
-  private int baudRate;
+  private final Context context;
+  private final int baudRate;
   private boolean busy;
   private int vendorId;
   private int productId;
@@ -48,102 +44,93 @@ public class UsbConnection {
   private String deviceName;
   private String manufacturerName;
 
-
   public UsbConnection(Context context, int baudRate) {
     this.context = context;
     this.baudRate = baudRate;
-    mLocalBroadcastManager = LocalBroadcastManager.getInstance(this.context);
+    localBroadcastManager = LocalBroadcastManager.getInstance(this.context);
     usbManager = (UsbManager) context.getSystemService(Context.USB_SERVICE);
-    usbPermissionIntent = PendingIntent.getBroadcast(this.context, 0, new Intent(ACTION_USB_PERMISSION), 0);
+    usbPermissionIntent =
+        PendingIntent.getBroadcast(this.context, 0, new Intent(ACTION_USB_PERMISSION), 0);
   }
 
-  private UsbSerialInterface.UsbReadCallback callback = new UsbSerialInterface.UsbReadCallback() {
-    @Override
-    public void onReceivedData(byte[] data) {
-      try {
-        String dataUtf8 = new String(data, "UTF-8");
-        buffer += dataUtf8;
-        int index;
-        while ((index = buffer.indexOf('\n')) != -1) {
-          final String dataStr = buffer.substring(0, index).trim();
-          buffer = buffer.length() == index ? "" : buffer.substring(index + 1);
+  private final UsbSerialInterface.UsbReadCallback callback =
+      data -> {
+        try {
+          String dataUtf8 = new String(data, "UTF-8");
+          buffer += dataUtf8;
+          int index;
+          while ((index = buffer.indexOf('\n')) != -1) {
+            final String dataStr = buffer.substring(0, index).trim();
+            buffer = buffer.length() == index ? "" : buffer.substring(index + 1);
 
-          AsyncTask.execute(new Runnable() {
-            @Override
-            public void run() {
-              onSerialDataReceived(dataStr);
-            }
-          });
+            AsyncTask.execute(() -> onSerialDataReceived(dataStr));
+          }
+        } catch (UnsupportedEncodingException e) {
+          LOGGER.e("Error receiving USB data");
         }
-      } catch (UnsupportedEncodingException e) {
-        LOGGER.e("Error receiving USB data");
-      }
-    }
-  };
+      };
 
-
-  private final BroadcastReceiver usbPermissionReceiver = new BroadcastReceiver() {
-    public void onReceive(Context context, Intent intent) {
-      String action = intent.getAction();
-      if (ACTION_USB_PERMISSION.equals(action)) {
-        synchronized (this) {
-          UsbDevice usbDevice = (UsbDevice) intent.getParcelableExtra(UsbManager.EXTRA_DEVICE);
-          if (intent.getBooleanExtra(UsbManager.EXTRA_PERMISSION_GRANTED, false)) {
-            if(usbDevice != null){
-              //call method to set up device communication
-              startSerialConnection(usbDevice);
+  private final BroadcastReceiver usbPermissionReceiver =
+      new BroadcastReceiver() {
+        public void onReceive(Context context, Intent intent) {
+          String action = intent.getAction();
+          if (ACTION_USB_PERMISSION.equals(action)) {
+            synchronized (this) {
+              UsbDevice usbDevice = intent.getParcelableExtra(UsbManager.EXTRA_DEVICE);
+              if (intent.getBooleanExtra(UsbManager.EXTRA_PERMISSION_GRANTED, false)) {
+                if (usbDevice != null) {
+                  // call method to set up device communication
+                  startSerialConnection(usbDevice);
+                }
+              } else {
+                LOGGER.d("Permission denied for device " + usbDevice);
+                Toast.makeText(
+                        UsbConnection.this.context,
+                        "USB Host permission is required!",
+                        Toast.LENGTH_LONG)
+                    .show();
+              }
             }
           }
-          else {
-            LOGGER.d("Permission denied for device " + usbDevice);
-            Toast.makeText(
-                    UsbConnection.this.context,
-                    "USB Host permission is required!",
-                    Toast.LENGTH_LONG).show();
+        }
+      };
+
+  private final BroadcastReceiver usbDetachedReceiver =
+      new BroadcastReceiver() {
+        public void onReceive(Context context, Intent intent) {
+          String action = intent.getAction();
+
+          if (UsbManager.ACTION_USB_DEVICE_DETACHED.equals(action)) {
+            UsbDevice device = intent.getParcelableExtra(UsbManager.EXTRA_DEVICE);
+            if (device != null) {
+              LOGGER.i("USB device detached");
+              stopUsbConnection();
+            }
           }
         }
-      }
-    }
-  };
-
-  private final BroadcastReceiver usbDetachedReceiver = new BroadcastReceiver() {
-    public void onReceive(Context context, Intent intent) {
-      String action = intent.getAction();
-
-      if (UsbManager.ACTION_USB_DEVICE_DETACHED.equals(action)) {
-        UsbDevice device = intent.getParcelableExtra(UsbManager.EXTRA_DEVICE);
-        if (device != null) {
-          LOGGER.i("USB device detached");
-          stopUsbConnection();
-        }
-      }
-    }
-  };
-
+      };
 
   public boolean startUsbConnection() {
     IntentFilter usbPermissionFilter = new IntentFilter(ACTION_USB_PERMISSION);
-    mLocalBroadcastManager.registerReceiver(usbPermissionReceiver, usbPermissionFilter);
+    localBroadcastManager.registerReceiver(usbPermissionReceiver, usbPermissionFilter);
     // Detach events are sent as a system-wide broadcast
     IntentFilter usbDetachedFilter = new IntentFilter(UsbManager.ACTION_USB_DEVICE_DETACHED);
-    mLocalBroadcastManager.registerReceiver(usbDetachedReceiver, usbDetachedFilter);
+    localBroadcastManager.registerReceiver(usbDetachedReceiver, usbDetachedFilter);
 
     Map<String, UsbDevice> connectedDevices = usbManager.getDeviceList();
     if (!connectedDevices.isEmpty()) {
       for (UsbDevice usbDevice : connectedDevices.values()) {
-        //if (usbDevice.getVendorId() == USB_VENDOR_ID && usbDevice.getProductId() == USB_PRODUCT_ID) {
-          LOGGER.i("Device found: " + usbDevice.getDeviceName());
-          if (usbManager.hasPermission(usbDevice)) {
-            return startSerialConnection(usbDevice);
-          }
-          else {
-            usbManager.requestPermission(usbDevice, usbPermissionIntent);
-            Toast.makeText(context,
-                    "Please allow USB Host connection.",
-                    Toast.LENGTH_SHORT).show();
-            return false;
-          }
-        //}
+        // if (usbDevice.getVendorId() == USB_VENDOR_ID && usbDevice.getProductId() ==
+        // USB_PRODUCT_ID) {
+        LOGGER.i("Device found: " + usbDevice.getDeviceName());
+        if (usbManager.hasPermission(usbDevice)) {
+          return startSerialConnection(usbDevice);
+        } else {
+          usbManager.requestPermission(usbDevice, usbPermissionIntent);
+          Toast.makeText(context, "Please allow USB Host connection.", Toast.LENGTH_SHORT).show();
+          return false;
+        }
+        // }
       }
     }
     LOGGER.w("Could not start USB connection - No devices found");
@@ -182,7 +169,8 @@ public class UsbConnection {
   private void onSerialDataReceived(String data) {
     // Add whatever you want here
     LOGGER.i("Serial data received: " + data);
-    mLocalBroadcastManager.sendBroadcast(new Intent(SensorService.USB_ACTION_DATA_RECEIVED)
+    localBroadcastManager.sendBroadcast(
+        new Intent(Constants.USB_ACTION_DATA_RECEIVED)
             .putExtra("from", "usb")
             .putExtra("data", data));
   }
@@ -200,16 +188,17 @@ public class UsbConnection {
       serialDevice = null;
       connection = null;
     }
-    mLocalBroadcastManager.unregisterReceiver(usbPermissionReceiver);
-    mLocalBroadcastManager.unregisterReceiver(usbDetachedReceiver);
+    localBroadcastManager.unregisterReceiver(usbPermissionReceiver);
+    localBroadcastManager.unregisterReceiver(usbDetachedReceiver);
   }
 
   public void send(String msg) {
-    busy = true;
-    serialDevice.write(msg.getBytes(UTF_8));
-    busy = false;
+    if (isOpen() && !isBusy()) {
+      busy = true;
+      serialDevice.write(msg.getBytes(UTF_8));
+      busy = false;
+    }
   }
-
 
   public boolean isOpen() {
     return connection != null;
