@@ -1,4 +1,4 @@
-//Created by Matthias Mueller - Intel Intelligent Systems Lab - 2020
+// Created by Matthias Mueller - Intel Intelligent Systems Lab - 2020
 
 package org.openbot.tflite;
 
@@ -6,12 +6,10 @@ import android.app.Activity;
 import android.graphics.Bitmap;
 import android.os.SystemClock;
 import android.os.Trace;
-
-import org.openbot.ControlSignal;
-
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
+import org.openbot.env.Vehicle;
 
 public abstract class Autopilot extends Network {
 
@@ -28,26 +26,21 @@ public abstract class Autopilot extends Network {
   /** A ByteBuffer to hold image data, to be feed into Tensorflow Lite as inputs. */
   protected ByteBuffer indicatorBuffer = null;
 
-
-  public static Autopilot create (Activity activity, Model model, Device device, int numThreads)
-          throws IOException {
-    switch (model) {
-      case AUTOPILOT_F:
-        return new AutopilotFloat(activity, device, numThreads);
-      default:
-        return new AutopilotFloat(activity, device, numThreads);
-    }
+  public static Autopilot create(Activity activity, Model model, Device device, int numThreads)
+      throws IOException {
+    return new AutopilotFloat(activity, model, device, numThreads);
   }
 
   /** Initializes a {@code Autopilot}. */
-  protected Autopilot(Activity activity, Device device, int numThreads) throws IOException {
-    super(activity, device, numThreads);
+  protected Autopilot(Activity activity, Model model, Device device, int numThreads)
+      throws IOException {
+    super(activity, model, device, numThreads);
     indicatorBuffer = ByteBuffer.allocateDirect(4);
     indicatorBuffer.order(ByteOrder.nativeOrder());
     LOGGER.d("Created a Tensorflow Lite Autopilot.");
   }
 
-  private void convertIndicatorToByteBuffer (int indicator) {
+  private void convertIndicatorToByteBuffer(int indicator) {
     if (indicatorBuffer == null) {
       return;
     }
@@ -55,7 +48,7 @@ public abstract class Autopilot extends Network {
     indicatorBuffer.putFloat(indicator);
   }
 
-  public ControlSignal recognizeImage(final Bitmap bitmap, final int indicator) {
+  public Vehicle.Control recognizeImage(final Bitmap bitmap, final int indicator) {
     // Log this method so that it can be analyzed with systrace.
     Trace.beginSection("recognizeImage");
     Trace.beginSection("preprocessBitmap");
@@ -65,24 +58,22 @@ public abstract class Autopilot extends Network {
 
     // Run the inference call.
     Trace.beginSection("runInference");
-    long startTime = SystemClock.uptimeMillis();
+    long startTime = SystemClock.elapsedRealtime();
     Object[] inputArray;
     if (tflite.getInputIndex("cmd_input") == 0) {
-      inputArray = new Object[] {indicatorBuffer,imgData};
-    }
-    else {
-      inputArray = new Object[] {imgData,indicatorBuffer};
+      inputArray = new Object[] {indicatorBuffer, imgData};
+    } else {
+      inputArray = new Object[] {imgData, indicatorBuffer};
     }
 
     float[][] predicted_ctrl = new float[1][2];
     outputMap.put(0, predicted_ctrl);
     tflite.runForMultipleInputsOutputs(inputArray, outputMap);
-    long endTime = SystemClock.uptimeMillis();
+    long endTime = SystemClock.elapsedRealtime();
     Trace.endSection();
     LOGGER.v("Timecost to run model inference: " + (endTime - startTime));
 
     Trace.endSection(); // "recognizeImage"
-    return new ControlSignal(predicted_ctrl[0][0], predicted_ctrl[0][1]);
+    return new Vehicle.Control(predicted_ctrl[0][0], predicted_ctrl[0][1]);
   }
-
 }
