@@ -13,18 +13,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-//Modified by Matthias Mueller - Intel Intelligent Systems Lab - 2020
+// Modified by Matthias Mueller - Intel Intelligent Systems Lab - 2020
 
 package org.openbot.tflite;
+
 import android.app.Activity;
 import android.content.res.AssetFileDescriptor;
 import android.graphics.Bitmap;
 import android.graphics.RectF;
 import android.os.SystemClock;
-
-import org.tensorflow.lite.Interpreter;
-import org.tensorflow.lite.gpu.GpuDelegate;
-
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -33,29 +31,25 @@ import java.nio.MappedByteBuffer;
 import java.nio.channels.FileChannel;
 import java.util.HashMap;
 import java.util.Map;
-
+import org.jetbrains.annotations.NotNull;
 import org.openbot.env.Logger;
+import org.tensorflow.lite.Interpreter;
+import org.tensorflow.lite.gpu.GpuDelegate;
 
 public abstract class Network {
 
   protected static final Logger LOGGER = new Logger();
 
-  /** The model. */
-  public enum Model {
-    DETECTOR_V1_1_0_Q,
-    DETECTOR_V3_S_Q,
-    AUTOPILOT_F,
-  }
-
   /** The runtime device type used for execution. */
   public enum Device {
     CPU,
-    NNAPI,
-    GPU
+    GPU,
+    NNAPI
   }
 
   /** Dimensions of inputs. */
   protected static final int DIM_BATCH_SIZE = 1;
+
   protected static final int DIM_PIXEL_SIZE = 3;
 
   /** Preallocated buffers for storing image data in. */
@@ -63,9 +57,6 @@ public abstract class Network {
 
   /** Options for configuring the Interpreter. */
   protected final Interpreter.Options tfliteOptions = new Interpreter.Options();
-
-  /** The loaded TensorFlow Lite model. */
-  protected MappedByteBuffer tfliteModel;
 
   /** Optional GPU delegate for accleration. */
   protected GpuDelegate gpuDelegate = null;
@@ -79,9 +70,8 @@ public abstract class Network {
   protected Map<Integer, Object> outputMap = new HashMap<>();
 
   /** Initializes a {@code Network}. */
-  protected Network(Activity activity, Device device, int numThreads) throws IOException {
-
-    tfliteModel = loadModelFile(activity);
+  protected Network(Activity activity, Model model, Device device, int numThreads)
+      throws IOException {
     switch (device) {
       case NNAPI:
         tfliteOptions.setUseNNAPI(true);
@@ -94,16 +84,27 @@ public abstract class Network {
         break;
     }
     tfliteOptions.setNumThreads(numThreads);
-    tflite = new Interpreter(tfliteModel, tfliteOptions);
+    if (model.filename != null) {
+      File modelFile = getModelFile(activity, model);
+      tflite = new Interpreter(modelFile, tfliteOptions);
+    } else {
+      MappedByteBuffer tfliteModel = loadModelFile(activity);
+      tflite = new Interpreter(tfliteModel, tfliteOptions);
+    }
     imgData =
-            ByteBuffer.allocateDirect(
-                    DIM_BATCH_SIZE
-                            * getImageSizeX()
-                            * getImageSizeY()
-                            * DIM_PIXEL_SIZE
-                            * getNumBytesPerChannel());
+        ByteBuffer.allocateDirect(
+            DIM_BATCH_SIZE
+                * getImageSizeX()
+                * getImageSizeY()
+                * DIM_PIXEL_SIZE
+                * getNumBytesPerChannel());
     imgData.order(ByteOrder.nativeOrder());
     LOGGER.d("Created a Tensorflow Lite Network.");
+  }
+
+  @NotNull
+  private File getModelFile(Activity activity, Model model) {
+    return new File(activity.getFilesDir() + File.separator + model.filename);
   }
 
   /** Memory-map the model file in Assets. */
@@ -125,17 +126,16 @@ public abstract class Network {
     bitmap.getPixels(intValues, 0, bitmap.getWidth(), 0, 0, bitmap.getWidth(), bitmap.getHeight());
     // Convert the image to floating point.
     int pixel = 0;
-    long startTime = SystemClock.uptimeMillis();
+    long startTime = SystemClock.elapsedRealtime();
     for (int i = 0; i < getImageSizeX(); ++i) {
       for (int j = 0; j < getImageSizeY(); ++j) {
         final int val = intValues[pixel++];
         addPixelValue(val);
       }
     }
-    long endTime = SystemClock.uptimeMillis();
+    long endTime = SystemClock.elapsedRealtime();
     LOGGER.v("Timecost to put values into ByteBuffer: " + (endTime - startTime));
   }
-
 
   /** Closes the interpreter and model to release resources. */
   public void close() {
@@ -147,7 +147,6 @@ public abstract class Network {
       gpuDelegate.close();
       gpuDelegate = null;
     }
-    tfliteModel = null;
   }
 
   /**
@@ -198,5 +197,4 @@ public abstract class Network {
    * @return
    */
   public abstract RectF getCropRect();
-
 }
